LayoutTests/imported/w3c/web-platform-tests/svg/animations/svglengthlist-animation-unitType-expected.txt

ABCD

FAIL Test change of unit type for SVGLengthList animation. assert_equals: expected 5 but got 2
